Module Picos.Computation

A cancelable computation

A computation simply holds the status, i.e.

running,

returned, or

canceled,

of some sort of computation.

As a hopefully helpful analogy, a computation is basically like a cancelable promise and a basic non-cancelable promise can be implemented trivially on top of a computation.

To define a computation, one first creates it and then arranges for the computation to be completed by returning a value through it or by canceling it with an exception at some point in the future. There are no restrictions on what it means for a computation to be running. The cancelation status of a computation can be checked explicitly. Outside observers can attach triggers to a computation to get a signal when the computation is completed or await the computation. Outside observers may also be selectively given the ability to cancel a computation.

Here is an example:

let computation =
  Computation.create ()
in
let computer =
  Domain.spawn @@ fun () ->
    let rec fib i =
      Computation.check computation;
      if i <= 1 then
        i
      else
        fib (i - 1) + fib (i - 2)
    in
    Computation.capture computation
      fib 10
in
let finally () =
  Domain.join computer
in
Fun.protect ~finally @@ fun () ->

let canceler =
  Domain.spawn @@ fun () ->
    Unix.sleepf 0.1;
    Computation.cancel computation
    @@ Exn_bt.get_callstack 2 Exit
in
let finally () =
  Domain.join canceler
in
Fun.protect ~finally @@ fun () ->

Computation.await computation

In this framework, a fiber is always associated with at least a single computation. However, it is possible for multiple fibers to share a single computation and it is also possible for a single fiber to perform multiple computations.

Computations are not hierarchical. In other words, computations do not directly implement structured concurrency. However, it is possible to propagate cancelation to implement structured concurrency on top of computations.

Operations on computations are either wait-free or lock-free and designed to avoid starvation and complete in amortized constant time. The properties of operations to complete a computation depend on the properties of actions attached to the triggers.

Interface for creating

type (!'a, -'allowed) t

Represents a cancelable computation.

val create : unit -> ('a, [ `Await | `Cancel | `Return ]) t

create () creates a new computation in the running state.

val finished : (unit, [ `Await | `Cancel | `Return ]) t

finished is a constant finished computation.

val return : ('a, [> `Return ]) t -> 'a -> unit

return computation value attempts to complete the computation with the specified value.

val finish : (unit, [> `Return ]) t -> unit

finish computation is equivalent to return computation ().

val capture : ('a, [> `Cancel | `Return ]) t -> ('b -> 'a) -> 'b -> unit

capture computation fn x calls fn x and tries to complete the computation with the value returned or the exception raised by the call.

val cancel_after : ('a, [> `Await | `Cancel ]) t -> seconds:float -> Exn_bt.t -> unit

cancel_after ~seconds computation exn_bt arranges to cancel the computation after the specified time with the specified exception and backtrace. Completion of the computation before the specified time effectively cancels the timeout.

The implementation will first try to perform Cancel_after and falls back to a default implementation using Threads if possible.

raises Invalid_argument
if seconds is negative.

Interface for canceling

type 'a as_cancelable = ('a, [ `Await | `Cancel ]) t

Synonym for a computation that only allows cancelation aside from await.

type -'allowed packed =

| Packed : ('a, 'allowed) t -> 'allowed packed

An existential wrapper for computations.

type packed_as_cancelable = [ `Await | `Cancel ] packed

Synonym for a packed computation that only allows cancelation aside from await.

val cancel : ('a, [> `Cancel ]) t -> Exn_bt.t -> unit

cancel computation exn_bt attempts to mark the computation as canceled with the specified exception and backtrace.

Interface for polling

val is_running : ('a, _) t -> bool

is_running computation determines whether the computation is in the running state meaning that it has not yet been completed.

val canceled : ('a, _) t -> Exn_bt.t option

canceled computation returns the exception that the computation has been canceled with or returns None in case the computation has not been canceled.

val check : ('a, _) t -> unit

check computation is equivalent to Option.iter Exn_bt.raise (canceled computation).

Interface for awaiting

val try_attach : ('a, [> `Await ]) t -> [> `Signal ] Trigger.t -> bool

try_attach computation trigger tries to attach the trigger to be signaled on completion of the computation and returns true on success. Otherwise returns false, which means that the computation has already been completed.

val detach : ('a, [> `Await ]) t -> [> `Signal ] Trigger.t -> unit

detach computation trigger signals the trigger and detaches it from the computation.

val await : ('a, [> `Await ]) t -> 'a

await computation waits for the computation to complete and either returns the value of the completed computation or raises the exception the computation was canceled with.

Interface for propagating cancelation

val canceler : from:('a, [> `Await ]) t -> into:('b, [> `Cancel ]) t -> [ `Signal ] Trigger.t

canceler ~from ~into creates a trigger that propagates cancelation from one computation into another on signal.

The returned trigger is usually attached to the computation from which cancelation is to be propagated and the trigger should usually also be detached after it is no longer needed.

Interface for schedulers

type Stdlib.Effect.t +=

| Cancel_after : {

seconds : float;

exn_bt : Exn_bt.t;

computation : 'a as_cancelable;

} -> unit Stdlib.Effect.t

Schedulers may handle the Cancel_after effect to customize the behavior of cancel_after.

Design rationale

The computation concept can be seen as a generalization of both a cancelation context or token and of a promise. Unlike a typical promise mechanism, a computation can be canceled. Unlike a typical cancelation mechanism, a computation can and should also be completed in case it is not canceled. This promotes proper scoping of computations and resource cleanup at completion, which is how the design evolved from a more traditional cancelation context design.

In this framework, every fiber has an associated computation. Being able to return a value through the computation means that no separate promise is necessarily required to hold the result of a fiber. On the other hand, in this framework, multiple fibers may share a single computation. This allows multiple fibers to be canceled efficiently through a single atomic update. In other words, the design allows various higher level patterns to be implemented efficiently.

Instead of directly implementing a hierarchy of computations, the design allows attaching triggers to computations and a special trigger constructor is provided for propagating cancelation. This helps to keep the implementation lean, i.e. not substantially heavier than a typical promise implementation.

Finally, just like with Trigger.Await, a key idea is that the handler of Computation.Cancel_after does not need to run arbitrary user defined code. The action of any trigger attached to a computation either comes from some scheduler calling Trigger.on_signal or from Computation.canceler.