Module Picos.Fiber

An independent thread of execution

A fiber corresponds to an independent thread of execution. Fibers are created by schedulers in response to Spawn effects. A fiber is associated with a computation and either forbids or permits the scheduler from propagating cancelation to it. A fiber also has an associated fiber local storage.

Interface for rescheduling

val yield : unit -> unit

yield () asks the current fiber to be re-scheduled.

On OCaml 5, yield will first attempt to perform the Yield effect and falls back to the OCaml 4 default implementation calling Thread.yield () when available.

Interface for spawning

val spawn : forbid:bool -> ('a, [> `Await | `Cancel ]) Computation.t -> (unit -> unit) list -> unit

spawn ~forbid computation mains starts new fibers by performing the Spawn effect. The fibers will share the same computation and start with propagation of cancelation forbidden or permitted depending on the forbid flag.

Note that any computation, including the computation of the current fiber, may be passed as the computation for new fibers. Higher level libraries are free to implement the desired structuring principles.

⚠️ Behavior is undefined if any function in mains raises an exception. For example, raising an exception might terminate the whole application (recommended, but not required) or the exception might be ignored. In other words, the caller must arrange for the computation to be completed and errors reported in a desired manner.

On OCaml 5, spawn will first try to perform the Spawn effect and falls back to the OCaml 4 default implementation creating Threads when available.

Interface for current fiber

type -'allowed t

Represents a fiber.

val current : unit -> [ `Sync | `Async ] t

current () returns the current fiber.

On OCaml 5, current will first try to perform the Current effect and falls back to the OCaml 4 default implementation using Thread-local storage.

⚠️ Operations that require the fiber handle to be [> `Sync] can only be called safely from the fiber itself.

val has_forbidden : [> `Sync ] t -> bool

has_forbidden fiber determines whether the fiber forbids or permits the scheduler from propagating cancelation to it.

val forbid : [> `Sync ] t -> (unit -> 'a) -> 'a

forbid fiber thunk tells the scheduler that cancelation must not be propagated to the fiber during the execution of thunk.

Note that this does not prevent the associated computation from being canceled. This only tells the scheduler not to propagate cancelation to the fiber.

val permit : [> `Sync ] t -> (unit -> 'a) -> 'a

permit fiber thunk tells the scheduler that cancelation may be propagated to the fiber during the execution of thunk.

val canceled : [> `Sync ] t -> Exn_bt.t option

canceled fiber is equivalent to:

if Fiber.has_forbidden fiber then
  None
else
  let (Packed computation) =
    Fiber.computation fiber
  in
  Computation.canceled computation

val check : [> `Sync ] t -> unit

check fiber is equivalent to:

if not (Fiber.has_forbidden fiber) then
  let (Packed computation) =
    Fiber.computation fiber
  in
  Computation.check computation

module FLS : sig ... end

Fiber local storage

Interface for foreign fiber

type as_async = [ `Async ] t

Synonym for a fiber with the allowed operations restricted to those that can be performed from outside of the fiber.

val equal : 'allowed t -> 'allowed t -> bool

equal fiber1 fiber2 determines whether fiber1 and fiber2 are one and the same fiber.

val computation : [> `Async ] t -> Computation.packed_as_cancelable

computation fiber returns the computation that the fiber has been created with.

val try_attach : [> `Async ] t -> [> `Signal ] Trigger.t -> bool

try_attach fiber trigger is equivalent to let Packed c = computation fiber in Computation.try_attach c trigger.

val detach : [> `Async ] t -> [> `Signal ] Trigger.t -> unit

detach fiber trigger is equivalent to let Packed c = computation fiber in Computation.detach c trigger.

Interface for schedulers

val create : forbid:bool -> ('a, [> `Await | `Cancel ]) Computation.t -> [ `Sync | `Async ] t

create ~forbid computation creates a new fiber.

val continue : 'a t -> ('v, 'r) Stdlib.Effect.Deep.continuation -> 'v -> 'r

continue fiber k v is equivalent to:

match Fiber.canceled fiber with
| None -> Effect.Deep.continue k v
| Some exn_bt -> Exn_bt.discontinue k exn_bt

val continue_with : 'a t -> ('v, 'b) Stdlib.Effect.Shallow.continuation -> 'v -> ('b, 'r) Stdlib.Effect.Shallow.handler -> 'r

continue_with fiber k v h is equivalent to:

match Fiber.canceled fiber with
| None -> Effect.Shallow.continue_with k v h
| Some exn_bt -> Exn_bt.discontinue_with k exn_bt h

type Stdlib.Effect.t +=

| Current : [ `Sync | `Async ] t Stdlib.Effect.t

Schedulers may handle the Current effect to customize the behavior of current.

A handler should eventually either continue the fiber or discontinue it in case the fiber permits propagation of cancelation and the associated computation has been canceled.

Note that in typical use cases of current it makes sense to give priority to the fiber performing Current, but this is not required.

type Stdlib.Effect.t +=

| Yield : unit Stdlib.Effect.t

Schedulers may handle the Yield effect to customize the behavior of yield.

Just like with Current, a handler should either continue or discontinue the fiber, but, unlike with Current, the scheduler should give priority to running other ready fibers before resuming the fiber performing Yield.

type Stdlib.Effect.t +=

| Spawn : {

forbid : bool;

computation : 'a Computation.as_cancelable;

mains : (unit -> unit) list;

} -> unit Stdlib.Effect.t

Schedulers may handle the Spawn effect to customize the behavior of spawn.

The scheduler is free to run the newly created fibers on any domain and decide which fiber to give priority to.

Design rationale

The idea is that fibers correspond 1-to-1 with independent threads of execution. This allows a fiber to non-atomically store state related to a thread of execution.

The status of whether propagation of cancelation is forbidden or permitted could be stored in the fiber local storage. The justification for storing it directly with the fiber is that the implementation of some key synchronization and communication mechanisms, such as condition variables, requires the capability.

No integer fiber id is provided by default. It would seem that for most intents and purposes the identity of the fiber is sufficient. Fiber local storage can be used to implement a fiber id or e.g. a fiber hash.

The fiber local storage is designed for the purpose of extending fibers and to be as fast as possible. It is not intended for application programming.

Yield is provided as a separate effect to specifically communicate the intent that the current fiber should be rescheduled. This allows all the other effect handlers more freedom in choosing which fiber to schedule next.