Module Picos.Handler

Handler for the effects based operations of Picos for OCaml 4.

type 'c t = {

current : 'c -> Fiber.t;
(*
See Picos.Fiber.current.
*)

spawn : 'c -> Fiber.t -> (Fiber.t -> unit) -> unit;
(*
See Picos.Fiber.spawn.
*)

yield : 'c -> unit;
(*
See Picos.Fiber.yield.
*)

cancel_after : 'a. 'c -> 'a Computation.t -> seconds:float -> Exn_bt.t -> unit;
(*
See Picos.Computation.cancel_after.
*)

await : 'c -> Trigger.t -> Exn_bt.t option;
(*
See Picos.Trigger.await.
*)

}

A record of implementations of the primitive effects based operations of Picos. The operations take a context of type 'c as an argument.

val using : 'c t -> 'c -> (Fiber.t -> unit) -> unit

using handler context main sets the handler and the context for the handler of the primitive effects based operations of Picos while running main.

ℹ️ The behavior is that

on OCaml 4, using stores the handler in TLS, which allows the operations to be accessed during the execution of the thunk, and

on OCaml 5, using runs thunk with a deep effect handler that delegates to the operations of the handler.

⚠️ While this works on OCaml 5, you usually want to use a scheduler that implements an effect handler directly, because that is likely to perform better.