Module Picos.Trigger

Ability to await for a signal

To suspend and later resume the current thread of execution, one can create a trigger, arrange signal to be called on it, and await for the call.

Here is a simple example:

let trigger = Trigger.create () in

let signaler =
  Domain.spawn @@ fun () ->
    Trigger.signal trigger
in
let finally () =
  Domain.join signaler
in
Fun.protect ~finally @@ fun () ->

match Trigger.await trigger with
| None ->
  (* We were resumed normally. *)
  ()
| Some exn_bt ->
  (* We were canceled.

     Typically we'd cleanup here. *)
  Exn_bt.raise exn_bt

All operations on triggers are wait-free, with the obvious exception of await. The signal operation inherits the properties of the action attached with on_signal to the trigger.

Interface for suspending

type -'allowed t

Represents a trigger.

val create : unit -> [ `Await | `Signal ] t

create () allocates a new trigger in the initial state.

val is_initial : _ t -> bool

is_initial trigger determines whether the trigger is in the initial state.

⚠️ Consider using is_signaled instead of is_initial as in some contexts a trigger might reasonably be either in the initial or the awaiting state depending on the order in which things are being done.

val is_signaled : _ t -> bool

is_signaled trigger determines whether the trigger is in the signaled state.

val await : [> `Await ] t -> Exn_bt.t option

await trigger waits for the trigger to be signaled.

The return value is None in case the trigger was signaled before await or the fiber was resumed normally. Otherwise the return value is Some exn_bt, which indicates that the fiber has been canceled and the caller should raise the exception. In either case the caller is responsible for cleaning up. Usually this means making sure that no references to the trigger remain to avoid space leaks.

⚠️ Only the owner or creator of a trigger may call await and it is considered an error to make multiple concurrent calls to await.

On OCaml 5, await will first try to perform the Await effect and falls back to the OCaml 4 default implementation that suspends the underlying system level thread using a per thread Mutex and Condition variable.

raises Invalid_argument
if the trigger was in the awaiting state, which means that multiple concurrent calls of await are being made.

Interface for resuming

type as_signal = [ `Signal ] t

Synonym for a trigger that only allows the signal operation.

val signal : [> `Signal ] t -> unit

After signal trigger returns, the trigger has been put into the signaled state and any attached action has been called.

Note that under normal circumstances, signal should never raise an exception. If an exception is raised by signal, it means that the handler of Await has a bug or some catastrophic failure has occurred.

Interface for schedulers

val on_signal : [> `On | `Signal ] t -> 'x -> 'y -> ([> `Signal ] t -> 'x -> 'y -> unit) -> bool

on_signal trigger x y resume attempts to attach the resume action to the trigger and transition the trigger to the awaiting state. It must be safe to call resume trigger x y from any context that signal might be called from.

The return value is true in case the action was attached successfully. Otherwise the return value is false, which means that the trigger was already in the signaled state.

⚠️ The handler of Await should call on_signal at most once.

raises Invalid_argument
if the trigger was in the awaiting state, which means that either the owner or creator of the trigger made concurrent calls to await or the handler called on_signal more than once.

type Stdlib.Effect.t += private

| Await : [ `On | `Signal ] t -> Exn_bt.t option Stdlib.Effect.t

Schedulers may handle the Await effect to customize the behavior of await.

In case the fiber permits propagation of cancelation, the trigger should be attached to the computation of the fiber for the duration of suspending the fiber.

Whether being resumed due to cancelation or not, the trigger should be signaled before resuming the fiber.

The scheduler is free to choose which ready fiber to resume next.

Design rationale

A key idea behind this design is that the handler for Await does not need to run arbitrary user defined code while suspending a fiber: the handler calls on_signal by itself. This should make it easier to get both the handler and the user code correct.

Another key idea is that the signal operation provides no feedback as to the outcome regarding cancelation. Calling signal merely guarantees that the caller of await will return. This means that the point at which cancelation must be determined can be as late as possible. A scheduler can check the cancelation status just before calling continue and it is, of course, possible to check the cancelation status earlier. This allows maximal flexibility for the handler of Await.

The consequence of this is that the only place to handle cancelation is at the point of await. This makes the design simpler and should make it easier for the user to get the handling of cancelation right. A minor detail is that await returns an option instead of raising an exception. The reason for this is that matching against an option is slightly faster than setting up an exception handler. Returning an option also clearly communicates the two different cases to handle.

On the other hand, the trigger mechanism does not have a way to specify a user-defined callback to perform cancelation immediately before the fiber is resumed. Such an immediately called callback could be useful for e.g. canceling an underlying IO request. One justification for not having such a callback is that cancelation is allowed to take place from outside of the scheduler, i.e. from another system level thread, and, in such a case, the callback could not be called immediately. Instead, the scheduler is free to choose how to schedule canceled and continued fibers and, assuming that fibers can be trusted, a scheduler may give priority to canceled fibers.

This design also separates the allocation of the atomic state for the trigger, or create, from await, and allows the state to be polled using is_signaled before calling await. This is particularly useful when the trigger might need to be inserted to multiple places and be signaled in parallel before the call of await.

No mechanism is provided to communicate any result with the signal. That can be done outside of the mechanism and is often not needed. This simplifies the design.

Once signal has been called, a trigger no longer refers to any other object and takes just two words of memory. This e.g. allows lazy removal of triggers.

To further understand the problem domain, in this design, in a suspend-resume scenario, there are three distinct pieces of state:

The state of shared data structure(s) used for communication and / or synchronization.

The state of the trigger.

The cancelation status of the fiber.

The trigger and cancelation status are both updated independently and atomically through code in this framework. The key requirement left for the user is to make sure that the state of the shared data structure is updated correctly independently of what await returns. So, for example, a mutex implementation must check, after getting Some exn_bt, what the state of the mutex is and how it should be updated.