Module type Picos.Implementation

Signature for an implementation of the primitive effects based operations of Picos.

ℹ️ On OCaml 4 most of the effects based operations will call Fiber.current implicitly to check for cancelation before finally calling the set implementation.

module Fiber : sig ... end

Implements primitive effects based operations of Picos.Fiber.

module Computation : sig ... end

Implements primitive effects based operations of Picos.Computation.

module Trigger : sig ... end

Implements primitive effects based operations of Picos.Trigger.