Module Picos_bootstrap.Computation

val error_negative_or_nan : unit -> 'a

val error_returned : unit -> 'a

type tx_state =

| Stopped
(*
= false
*)

| Started

| Aborted

type _ tx =

| Stopped : [> `Stopped ] tx

| Running : {

state : tx_state Stdlib.Atomic.t;

mutable completions : [ `Nil | `Completion ] completions;

} -> [> `Running ] tx

and _ completions =

| Unused : [> `Unused ] completions
(*
Affects Nils value!
*)

| Nil : [> `Nil ] completions
(*
= true
*)

| Completion : {

computation : 'a t;

before : ('a, [ `Continue ]) st;

completions : [ `Nil | `Completion ] completions;

} -> [> `Completion ] completions

and ('a, _) st =

| Canceled : {

exn : exn;

bt : Stdlib.Printexc.raw_backtrace;

mutable tx : [ `Stopped | `Running ] tx;

} -> ('a, [> `Canceled ]) st

| Returned : {

value : 'a;

mutable tx : [ `Stopped | `Running ] tx;

} -> ('a, [> `Returned ]) st

| Continue : {

balance_and_mode : int;

triggers : Trigger.t list;

} -> ('a, [> `Continue ]) st

and 'a state =

| S : ('a, [< `Canceled | `Returned | `Continue ]) st -> 'a state

and 'a t = 'a state Stdlib.Atomic.t

val fifo_bit : int

val one : int

val signal : ('a, [ `Continue ]) st -> unit

module Tx : sig ... end

val empty_fifo : 'a state

val empty_lifo : 'a state

val create : ?mode:[ `FIFO | `LIFO ] -> unit -> 'a state Stdlib.Atomic.t

val with_action : ?mode:[ `FIFO | `LIFO ] -> 'a -> 'b -> (Trigger.t -> 'a -> 'b -> unit) -> 'c state Stdlib.Atomic.t

val is_canceled : 'a state Stdlib.Atomic.t -> bool

val canceled : 'a state Stdlib.Atomic.t -> (exn * Stdlib.Printexc.raw_backtrace) option

val gc : int -> Trigger.t list -> Trigger.state Stdlib.Atomic.t list -> 'a state

gc is called when balance becomes negative by both try_attach and detach. This ensures that the O(n) lazy removal done by gc cannot cause starvation, because the only reason that CAS fails after gc is that someone else completed the gc.

val try_attach : 'a state Stdlib.Atomic.t -> Trigger.state Stdlib.Atomic.t -> bool

val unsafe_unsuspend : 'a state Stdlib.Atomic.t -> Backoff.t -> bool

val detach : 'a state Stdlib.Atomic.t -> Trigger.state Stdlib.Atomic.t -> unit

type packed =

| Packed : 'a t -> packed

This cannot be @@unboxed because Atomic.t is opaque

val is_running : 'a state Stdlib.Atomic.t -> bool

val try_terminate : 'a state Stdlib.Atomic.t -> 'a state -> Backoff.t -> bool

val returned_unit : 'a

val make_returned : 'a -> 'a state

val returned : 'a -> 'a state Stdlib.Atomic.t

val finished : unit state Stdlib.Atomic.t

val try_return : 'a state Stdlib.Atomic.t -> 'a -> bool

val try_finish : 'a state Stdlib.Atomic.t -> bool

val try_cancel : 'a state Stdlib.Atomic.t -> exn -> Stdlib.Printexc.raw_backtrace -> bool

val return : 'a state Stdlib.Atomic.t -> 'a -> unit

val finish : 'a state Stdlib.Atomic.t -> unit

val cancel : 'a state Stdlib.Atomic.t -> exn -> Stdlib.Printexc.raw_backtrace -> unit

val try_capture : 'a state Stdlib.Atomic.t -> ('b -> 'a) -> 'b -> bool

val capture : 'a state Stdlib.Atomic.t -> ('b -> 'a) -> 'b -> unit

val check : 'a state Stdlib.Atomic.t -> unit

exception Running

val peek_exn : 'a state Stdlib.Atomic.t -> 'a

val peek : 'a state Stdlib.Atomic.t -> ('a, exn * Stdlib.Printexc.raw_backtrace) Stdlib.result option

val propagate : 'a -> 'b state Stdlib.Atomic.t -> 'c state Stdlib.Atomic.t -> unit

val canceler : from:'a state Stdlib.Atomic.t -> into:'b state Stdlib.Atomic.t -> Trigger.state Stdlib.Atomic.t

val check_non_negative : float -> unit

val get_or : ('a state Stdlib.Atomic.t -> 'a state Stdlib.Atomic.t) -> 'a state Stdlib.Atomic.t -> 'a

val attach_canceler : from:'a state Stdlib.Atomic.t -> into:'b state Stdlib.Atomic.t -> Trigger.state Stdlib.Atomic.t