Module type Picos_lwt.Sleep

Minimal signature for an implementation of sleep using Lwt.

val sleep : float -> unit Lwt.t

sleep seconds should return a cancelable promise that resolves after given number of seconds (unless canceled).