Module Picos_structured

Basic structured concurrency primitives for Picos.

This library essentially provides one user level interface for structuring fibers with any Picos compatible scheduler. This library is both meant to serve as an example of what can be done and to also provide practical means for programming with fibers. Hopefully there will be many more libraries implemented in Picos like this providing different approaches, patterns, and idioms for structuring concurrent programs.

Modules

module Finally : sig ... end

Syntax for avoiding resource leaks.

module Control : sig ... end

Basic control operations and exceptions for structured concurrency.

module Bundle : sig ... end

A dynamic bundle of fibers guaranteed to be joined at the end.

module Run : sig ... end

Operations for running fibers in specific patterns.

Examples

First we open some modules for convenience:

open Picos_structured.Finally
open Picos_structured
open Picos_stdio
open Picos_sync

A simple echo server and clients

Let's build a simple TCP echo server and run it with some clients.

We first define a function for the server:

let server socket =
  Unix.listen socket 8;

  Bundle.join_after begin fun bundle ->
    while true do
      let^ client =
        finally Unix.close @@ fun () ->
        Unix.accept
          ~cloexec:true socket |> fst
      in

      (* Fork a fiber for client *)
      Bundle.fork bundle begin fun () ->
        let@ client = move client in
        Unix.set_nonblock client;

        let bytes =
          Bytes.create 100
        in
        let n =
          Unix.read client bytes 0
            (Bytes.length bytes)
        in
        Unix.write client bytes 0 n
        |> ignore
      end
    done
  end

The server function expects a bound socket and starts listening. For each accepted client the server forks a new fiber to handle it.

Let's then define a function for the clients:

let client addr =
  let@ socket =
    finally Unix.close @@ fun () ->
    Unix.socket ~cloexec:true
      PF_INET SOCK_STREAM 0
  in
  Unix.set_nonblock socket;
  Unix.connect socket addr;

  let msg = "Hello!" in
  Unix.write_substring
    socket msg 0 (String.length msg)
  |> ignore;

  let bytes =
    Bytes.create (String.length msg)
  in
  let n =
    Unix.read socket bytes 0
      (Bytes.length bytes)
  in

  Printf.printf "Received: %s\n%!"
    (Bytes.sub_string bytes 0 n)

The client function takes the address of the server and connects a socket to the server address. It then writes a message to server and reads a reply from the server and prints it.

Here is the main program:

let main () =
  let@ socket =
    finally Unix.close @@ fun () ->
    Unix.socket ~cloexec:true
      PF_INET SOCK_STREAM 0
  in
  Unix.set_nonblock socket;

  (* Let system determine the port *)
  Unix.bind socket Unix.(
    ADDR_INET(inet_addr_loopback, 0));

  let addr =
    Unix.getsockname socket
  in

  Bundle.join_after begin fun bundle ->
    (* Start server *)
    Bundle.fork bundle (fun () ->
      server socket);

    (* Run clients concurrently *)
    Bundle.join_after begin fun bundle ->
      for _=1 to 5 do
        Bundle.fork bundle (fun () ->
          client addr)
      done
    end;

    (* Stop server *)
    Bundle.terminate bundle
  end

The main program creates a socket for the server and binds it. The server is then started as a new fiber. Then the clients are started to run concurrently. Finally the server is terminated.

Finally we run the main program with a scheduler:

# Picos_fifos.run ~forbid:false main
Received: Hello!
Received: Hello!
Received: Hello!
Received: Hello!
Received: Hello!
- : unit = ()

As an exercise, you might want to refactor the server to avoid moving the file descriptors and use a recursive accept loop instead.