Module Picos_sync

Basic communication and synchronization primitives for Picos.

The optional padded argument taken by several constructor functions, e.g. Mutex.create and Condition.create, defaults to false. When explicitly specified as ~padded:true the object is allocated in a way to avoid false sharing. For relatively long lived objects this can improve performance and make performance more stable at the cost of using more memory. It is not recommended to use ~padded:true for short lived objects.

module Mutex : sig ... end

A mutex implementation for Picos.

module Condition : sig ... end

A condition implementation for Picos.

module Lazy : sig ... end

A lazy implementation for Picos.