Auxiliary libraries for Picos

This package contains auxiliary libraries used in the implementation of other Picos libraries.

Picos_aux_htbl Lock-free hash table.

Picos_aux_mpmcq Lock-free multi-producer, multi-consumer queue.

Picos_aux_mpscq Lock-free multi-producer, single-consumer queue.

Picos_aux_rc External reference counting tables for disposable resources.

Some of these libraries might be moved to other packages in the future.