Asynchronous IO system for Picos

This package provides an asynchronous IO system for Picos.

Picos_stdio Basic IO facilities based on OCaml standard libraries for Picos.

Picos_stdio_select Basic Unix.select based IO event loop for Picos.

Picos_stdio_fd Externally reference counted file descriptors.

The IO libraries in this package are currently built only on top of the standard libraries distributed with OCaml and are hopefully useful for building simple applications.