Lwt interface for Picos

This package provides an alternative direct style interface to programming with Lwt via Picos.

Picos_lwt Direct style Picos compatible interface to Lwt for OCaml 5.

Picos_lwt_unix Direct style Picos compatible interface to Lwt with Lwt_unix for OCaml 5.