Module Fiber.FLS

Fiber local storage

Fiber local storage is intended for use as a low overhead storage mechanism for fiber extensions. For example, one might associate a priority value with each fiber for a scheduler that uses a priority queue or one might use FLS to store unique id values for fibers.

Here is an example of how one might define a fiber id:

let fiber_id_key =
  let next = Atomic.make 0 in
  let key =
    Fiber.FLS.new_key
    @@ Computed (fun () ->
        Atomic.fetch_and_add next 1)
  in
  (key :> _ Fiber.FLS.as_read_only)

Here is an example of how one might define a fiber priority:

let fiber_priority_key =
  Fiber.FLS.new_key (Constant 0)

A priority based scheduler could then use get to access the priority of a fiber highly efficiently. Fibers that are fine with the default priority do not necessarily need to store a priority at all.

type (!'a, -'allowed) key

Represents a key for storing values of type 'a in storage associated with fibers.

type 'a as_read_only = ('a, [ `Get ]) key

Synonym for a key with allowed operations restricted to reading.

type 'a initial =

| Constant of 'a

| Computed of unit -> 'a

Type to specify initial values for fibers.

val new_key : 'a initial -> ('a, [ `Get | `Set ]) key

new_key initial allocates a new key for associating values in storage associated with fibers. The initial value for every fiber is either the given Constant or is Computed with the given function. If the initial value is a constant, no value needs to be stored unless the value is explicitly updated.

⚠️ New keys should not be created dynamically.

⚠️ Operations that require the fiber handle to be [> `Sync] can only be called safely from the fiber itself.

val get : [> `Sync ] t -> ('a, [> `Get ]) key -> 'a

get fiber key returns the value associated with the key in the storage associated with the fiber.

val set : [> `Sync ] t -> ('a, [> `Set ]) key -> 'a -> unit

set fiber key value sets the value associated with the key to the given value in the storage associated with the fiber.