Module Picos_lwt

A functor for building a Picos compatible direct style interface to Lwt with given implementation of sleep.

This basically gives you an alternative direct style interface to programming with Lwt. All the scheduling decisions will be made by Lwt.

module type Sleep = sig ... end

Minimal signature for an implementation of sleep using Lwt.

module type S = sig ... end

Direct style Picos compatible interface to Lwt.

module Make (_ : Sleep) : S

Make (Sleep) creates a Picos compatible interface to Lwt with given implementation of sleep.