Module Picos_structured.Run

Operations for running fibers in specific patterns.

val all : (unit -> unit) list -> unit

all actions starts all the actions as separate fibers and waits until they all complete.

val any : (unit -> unit) list -> unit

any actions starts all the actions as separate fibers and waits until at least one of them completes. The rest of the started fibers will then be canceled.

⚠️ Calling any [] is equivalent to calling block ().