Module Picos_threaded

A basic Thread based implementation of the effects based operations of Picos.

Briefly:

current returns the current fiber from the per thread state.

spawn creates a new Thread for each fiber.

yield just calls Thread.yield.

cancel_after uses DLS to store a priority queue of timeouts and a per-domain background timeout thread that runs a Unix.select loop to cancel computations.

await uses a per thread mutex and condition to suspend the thread.

⚠️ This implementation is probably suitable for simple applications that do not spawn a lot of fibers. If an application uses a lot of short lived fibers, then a more sophisticated implementation using some sort of thread pool will likely perform significantly better.

type t

The per thread state for a handler.

val create : forbid:bool -> 'a Picos.Computation.t -> t

create ~forbid computation creates a new per thread state for a fiber with the given computation and propagation of cancelation forbidden as given.

val handler : t Picos.Handler.t

A basic handler of the effects based operations of Picos using Threads and a Unix.select based IO event loop.