Sample schedulers for Picos

This package contains schedulers provided as samples for Picos.

Picos_mux_fifo Basic single-threaded effects based Picos compatible scheduler for OCaml 5.

Picos_mux_multififo Basic multi-threaded effects based Picos compatible scheduler for OCaml 5.

Picos_mux_random Randomized multi-threaded effects based Picos compatible scheduler for OCaml 5.

Picos_mux_thread Basic Thread based Picos compatible scheduler for OCaml 4.

You may find these useful for both understanding the core Picos interface and for testing your own libraries implemented in Picos.